custom-elements/instantiating-custom-elements/extensions-to-document-interface/create-element-namespace.html

<!DOCTYPE html>
<html>
<head>
<title>Namespace for createElement is HTML Namespace</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#extensions-to-document-interface-to-instantiate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    var name = 'x-a';
    
    doc.register(name);

    var customElement = doc.createElement(name);

    assert_equals(customElement.namespaceURI, HTML_NAMESPACE, 
            'Wrong custom element namespace');

}, 'Namespace for createElement is HTML Namespace');


test(function() {
    
    for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
        var element = HTML5_ELEMENTS[i];
        
        var doc = newHTMLDocument();
        
        var obj = doc.createElement(element);
        
        var name = 'x-2-' + element;
        
        var proto = Object.create(obj.constructor.prototype);
        
        doc.register(name, {
            prototype: Object.create(proto),
            extends: element
        });
        
        var customElement = doc.createElement(element, name);
            
        assert_equals(customElement.namespaceURI, HTML_NAMESPACE, 
                'Wrong custom element namespace for element extending ' + element);
    }

}, 'Namespace for createElement is HTML Namespace. Type extension specified');

</script>
</body>
</html>
